Intent: clarify cache responsibilities Operators involved: [Self-organization, Coupling] Affected invariants: [#7, #8] Key changes: - document legacy cache shim - annotate edge cache for test monkeypatching Expected risks/dissonances: legacy shim may persist; marked for removal Metrics: (C(t), Si, νf, phase) unchanged Equivalence map: tnfr.helpers.cache -> tnfr.helpers.node_cache & tnfr.helpers.edge_cache
