+++ tests/test_node_offset.py
@@ -1,7 +1,7 @@
 """Pruebas del offset de ``NodoTNFR``."""
 
 from tnfr.node import NodoTNFR, NodoNX
-from tnfr.helpers.cache import ensure_node_offset_map
+from tnfr.helpers import ensure_node_offset_map
 
 
 def test_offset_non_zero_with_mapping(graph_canon):
