+++ tests/test_prepare_dnfr_data_performance.py
@@ -4,7 +4,7 @@ import pytest
 
 from tnfr.constants import get_aliases
 from tnfr.dynamics import _prepare_dnfr_data
-from tnfr.helpers.cache import cached_nodes_and_A
+from tnfr.helpers import cached_nodes_and_A
 from tnfr.alias import set_attr, collect_attr
 
 ALIAS_THETA = get_aliases("THETA")
